requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.21.0
bitsandbytes>=0.41.0
pandas>=1.5.3
numpy>=1.24.2
scipy>=1.10.1
matplotlib>=3.7.1
seaborn>=0.12.2
tqdm>=4.65.0
wandb>=0.15.0
sentencepiece>=0.1.99
protobuf>=4.25.1
einops>=0.7.0
scikit-learn>=1.3.0
biopython>=1.81
requests>=2.28.2
